Доработки по дз
